Adding Getting One And All works

+++ SEM2/Exam/collectionOperations.js
@@ -1,14 +1,24 @@
 
+const AdObject = require('./AdvertiseController/AdObject');
 const auth = require('./authMiddleware');
 const client = require('./mongoDBCliend').getDb;
 const errorLog = require('./errorLog').logError;
+
+function updateElement(basicElement, elementToUpdate) {
+    basicElement.title = elementToUpdate.title ?? basicElement.title;
+    basicElement.description = elementToUpdate.description ?? basicElement.description;
+    basicElement.category = elementToUpdate.category ?? basicElement.category;
+    basicElement.tagArray = elementToUpdate.tagArray ?? basicElement.tagArray;
+    basicElement.price = elementToUpdate.price ?? basicElement.price;
+}
+
 
 async function addElementToCollection(collectionName, element) {
     try {
         const collection = client().collection(collectionName);
         const result = await collection.insertOne(element);
         if (!result) {
-            throw '500 - Internal Server Error'
+            throw '500 - Internal Server Error';
         }
         return result;
     } catch (error) {
@@ -19,9 +29,14 @@ async function addElementToCollection(collectionName, element) {
 async function deleteElementFromCollection(collectionName, elementId) {
     try {
         const collection = client().collection(collectionName);
-        const element = collection.find(x => x.Id === elementId);
+        let element = await collection.find().toArray();
+        element = element.find(x => x.toString().includes(elementId));
+
         if (!auth.isUserCanModify(element.CreatorId)) {
             throw '401 - unauthorized';
+        }
+        if (!element) {
+            throw '404 - Element not found'
         }
         const result = await collection.deleteOne(element);
         return result;
@@ -30,14 +45,18 @@ async function deleteElementFromCollection(collectionName, elementId) {
     }
 }
 
-async function updateElementInCollection(collectionName, element) {
+async function updateElementInCollection(collectionName, elementToUpdate) {
     try {
         const collection = client().collection(collectionName);
         const options = { upsert: true };
-        const element = collection.find(x => x.Id === elementId);
+        const element = collection.find(x => x._id === elementToUpdate.id);
         if (!auth.isUserCanModify(element.CreatorId)) {
             throw '401 - unauthorized';
         }
+        if (!element) {
+            throw '404 - Element not found'
+        }
+        updateElement(element, elementToUpdate);
         const result = await collection.updateOne(element, options);
         return result;
     } catch (error) {
@@ -45,20 +64,28 @@ async function updateElementInCollection(collectionName, element) {
     }
 }
 async function getAllElements(collectionName) {
-
     try {
         const collection = client().collection(collectionName);
         const result = await collection.find().toArray();
+        if (!result) {
+            throw '404 - Element not found';
+        }
         return result;
     } catch (error) {
         errorLog("Error while getting all elements!\n " + error);
     }
 }
 
-async function getOneElement(collectionName, element) {
+async function getOneElement(collectionName, elementId) {
     try {
         const collection = client().collection(collectionName);
-        const result = await collection.find({ isActive: true }).toArray();
+
+        let result = await collection.find().toArray();
+        result = result.find(x => x.toString().includes(elementId));
+
+        if (!result) {
+            throw '404 - Element not found';
+        }
         return result;
     } catch (error) {
         errorLog("Error while getting one element!\n " + error);
